Avoid downloading description tab from wizard when opening front page
